+++ flutter_inappwebview_windows/windows/utils/util.h
@@ -1,13 +1,14 @@
 #ifndef FLUTTER_INAPPWEBVIEW_PLUGIN_UTIL_H_
 #define FLUTTER_INAPPWEBVIEW_PLUGIN_UTIL_H_
 
-#pragma comment(lib, "rpcrt4.lib")  // UuidCreate - Minimum supported OS Win 2000
-
 #include <optional>
 #include <rpc.h>
 #include <string>
 #include <variant>
 #include <Windows.h>
+#include <winrt/Windows.Foundation.h>
+
+#include "string.h"
 
 namespace flutter_inappwebview_plugin
 {
@@ -51,6 +52,32 @@ namespace flutter_inappwebview_plugin
     }
 
     return guid;
+  }
+
+  static inline std::string get_origin_from_url(const std::string& url)
+  {
+    try {
+      winrt::Windows::Foundation::Uri const uri{ utf8_to_wide(url) };
+      auto scheme = winrt::to_string(uri.SchemeName());
+      auto host = winrt::to_string(uri.Host());
+      auto uriPort = uri.Port();
+      std::string port = "";
+      if (uriPort > 0 && ((string_equals(scheme, "http") && uriPort != 80) || (string_equals(scheme, "https") && uriPort != 443))) {
+        port = ":" + std::to_string(uriPort);
+      }
+      return scheme + "://" + host + port;
+    }
+    catch (...) {
+      auto urlSplit = split(url, std::string{ "://" });
+      if (urlSplit.size() > 1) {
+        auto scheme = urlSplit[0];
+        auto afterScheme = urlSplit[1];
+        auto afterSchemeSplit = split(afterScheme, std::string{ "/" });
+        auto host = afterSchemeSplit[0];
+        return scheme + "://" + host;
+      }
+    }
+    return url;
   }
 }