+++ flutter_inappwebview_windows/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated flutter_inappwebview_platform_interface version to ^1.4.0
 - Updated `scrollMultiplier` default value from 6 to 1
 - Added support for `UserScript.allowedOriginRules` and `UserScript.forMainFrameOnly` parameters
+- Implemented `onReceivedHttpAuthRequest` WebView events
 - Fixed `get_optional_fl_map_value` implementation in `utils/flutter.h`
 
 ## 0.6.0